simplify typing of data passed in services

+++ Jacdac.Tests/PackTest.cs
@@ -25,8 +25,8 @@ namespace Jacdac.Tests
             var parsed = PacketEncoding.UnPack(format, buffer);
 
             Assert.True(parsed.Length == 1);
-            Assert.IsType<ushort>(parsed[0]);
-            Assert.Equal<ushort>(37778, (ushort)parsed[0]);
+            Assert.IsType<uint>(parsed[0]);
+            Assert.Equal<uint>(37778, (uint)parsed[0]);
         }
 
         [Fact]
@@ -37,8 +37,8 @@ namespace Jacdac.Tests
             var parsed = PacketEncoding.UnPack(format, buffer);
 
             Assert.True(parsed.Length == 1);
-            Assert.IsType<byte>(parsed[0]);
-            Assert.Equal<byte>(0x92, (byte)parsed[0]);
+            Assert.IsType<uint>(parsed[0]);
+            Assert.Equal<uint>(0x92, (uint)parsed[0]);
         }
 
         [Fact]
@@ -49,8 +49,8 @@ namespace Jacdac.Tests
             var parsed = PacketEncoding.UnPack(format, buffer);
 
             Assert.True(parsed.Length == 1);
-            Assert.IsType<UInt32>(parsed[0]);
-            Assert.Equal<UInt32>(4242, (uint)parsed[0]);
+            Assert.IsType<uint>(parsed[0]);
+            Assert.Equal<uint>(4242, (uint)parsed[0]);
         }
 
         [Fact]
@@ -110,9 +110,9 @@ namespace Jacdac.Tests
 
             Assert.True(parsed.Length == 2);
             Assert.IsType<string>(parsed[0]);
-            Assert.IsType<byte>(parsed[1]);
+            Assert.IsType<uint>(parsed[1]);
             Assert.Equal("Hello World", parsed[0]);
-            Assert.Equal<byte>(255, (byte)parsed[1]);
+            Assert.Equal<uint>(255, (uint)parsed[1]);
         }
 
         [Fact]
@@ -123,13 +123,13 @@ namespace Jacdac.Tests
             var parsed = PacketEncoding.UnPack(format, buffer);
 
             Assert.True(parsed.Length == 2);
-            Assert.IsType<byte>(parsed[0]);
-            Assert.Equal<byte>(127, (byte)parsed[0]);
+            Assert.IsType<uint>(parsed[0]);
+            Assert.Equal<uint>(127, (uint)parsed[0]);
 
             Assert.IsType<object[]>(parsed[1]);
-            Assert.Equal<ushort>(43981, (ushort)((object[])parsed[1])[0]);
-            Assert.Equal<ushort>(4660, (ushort)((object[])parsed[1])[1]);
-            Assert.Equal<ushort>(22136, (ushort)((object[])parsed[1])[2]);
+            Assert.Equal<uint>(43981, (uint)((object[])parsed[1])[0]);
+            Assert.Equal<uint>(4660, (uint)((object[])parsed[1])[1]);
+            Assert.Equal<uint>(22136, (uint)((object[])parsed[1])[2]);
         }
 
         [Fact]
@@ -140,51 +140,50 @@ namespace Jacdac.Tests
             var parsed = PacketEncoding.UnPack(format, buffer);
 
             Assert.True(parsed.Length == 2);
-            Assert.IsType<byte>(parsed[0]);
-            Assert.Equal<byte>(127, (byte)parsed[0]);
-
-            Assert.IsType<byte>(parsed[1]);
-            Assert.Equal<byte>(255, (byte)parsed[1]);
+            Assert.IsType<uint>(parsed[0]);
+            Assert.Equal<uint>(127, (uint)parsed[0]);
+            Assert.IsType<uint>(parsed[1]);
+            Assert.Equal<uint>(255, (uint)parsed[1]);
         }
 
         const int precision = 4;
 
         [Theory]
-        [InlineData("u8", new object[] { (byte)42 })]
-        [InlineData("u16", new object[] { (ushort)42 })]
-        [InlineData("u32", new object[] { (uint)42 })]
+        [InlineData("u8", new object[] { 42u })]
+        [InlineData("u16", new object[] { 42u })]
+        [InlineData("u32", new object[] { 42u })]
         [InlineData("u64", new object[] { (ulong)42 })]
-        [InlineData("i8", new object[] { (sbyte)-3 })]
-        [InlineData("i16", new object[] { (short)-42 })]
-        [InlineData("i32", new object[] { (int)-42 })]
+        [InlineData("i8", new object[] { -3 })]
+        [InlineData("i16", new object[] { -42 })]
+        [InlineData("i32", new object[] { -42 })]
         [InlineData("i64", new object[] { (long)-42 })]
-        [InlineData("u16 u16 i16", new object[] { (ushort)42, (ushort)77, (short)-10 })]
-        [InlineData("u16 z s", new object[] { (ushort)42, "foo", "bar" })]
-        [InlineData("u32 z s", new object[] { (uint)42, "foo", "bar" })]
-        [InlineData("i8 z s", new object[] { (sbyte)42, "foo", "bar" })]
-        [InlineData("u8 z s", new object[] { (byte)42, "foo12", "bar" })]
+        [InlineData("u16 u16 i16", new object[] { 42u, 77u, -10 })]
+        [InlineData("u16 z s", new object[] { 42u, "foo", "bar" })]
+        [InlineData("u32 z s", new object[] { 42u, "foo", "bar" })]
+        [InlineData("i8 z s", new object[] { 42, "foo", "bar" })]
+        [InlineData("u8 z s", new object[] { 42u, "foo12", "bar" })]
         [InlineData("u8 r: u8 z", new object[] {
-            (byte)42,
+            42u,
             new object[] {
-                new object[] { (byte)17, "xy" },
-                new object[] { (byte)18, "xx" }}
+                new object[] { 17u, "xy" },
+                new object[] { 18u, "xx" }}
             })]
         [InlineData("z b", new object[] { "foo12", new byte[] { 0, 1, 2, 3, 4 } })]
-        [InlineData("u16 r: u16", new object[] { (ushort)42, new object[] { new object[] { (ushort)17 }, new object[] { (ushort)18 } } })]
-        [InlineData("i8 s[9] u16 s[10] u8", new object[] { (sbyte)-100, "foo", (ushort)1000, "barbaz", (byte)250 })]
-        [InlineData("i8 x[4] s[9] u16 x[2] s[10] x[3] u8", new object[] { (sbyte)-100, "foo", (ushort)1000, "barbaz", (byte)250 })]
-        [InlineData("u16 u16[]", new object[] { (ushort)42, new object[] { (ushort)17, (ushort)18 } })]
-        [InlineData("u16 u16[]", new object[] { (ushort)42, new object[] { (ushort)18 } })]
-        [InlineData("u16 u16[]", new object[] { (ushort)42, new object[] { } })]
+        [InlineData("u16 r: u16", new object[] { 42u, new object[] { new object[] { 17u }, new object[] { 18u } } })]
+        [InlineData("i8 s[9] u16 s[10] u8", new object[] { -100, "foo", 1000u, "barbaz", 250u })]
+        [InlineData("i8 x[4] s[9] u16 x[2] s[10] x[3] u8", new object[] { -100, "foo", 1000u, "barbaz", 250u })]
+        [InlineData("u16 u16[]", new object[] { 42u, new object[] { 17u, 18u } })]
+        [InlineData("u16 u16[]", new object[] { 42u, new object[] { 18u } })]
+        [InlineData("u16 u16[]", new object[] { 42u, new object[] { } })]
         [InlineData("u0.16", new object[] { (float)0 }, precision)]
         [InlineData("u0.16", new object[] { (float)0.42 }, precision)]
         [InlineData("u0.16", new object[] { (float)1 }, precision)]
         [InlineData("i1.15", new object[] { (float)0 }, precision)]
         [InlineData("i1.15", new object[] { (float)1 }, precision)]
         [InlineData("i1.15", new object[] { (float)-1 }, precision)]
-        [InlineData("i16.16", new object[] { (float)0.1}, precision)]
+        [InlineData("i16.16", new object[] { (float)0.1 }, precision)]
         [InlineData("i16.16", new object[] { (float)1 }, precision)]
-        [InlineData("i16.16", new object[] { (float)Math.PI }, precision) ]
+        [InlineData("i16.16", new object[] { (float)Math.PI }, precision)]
         public void TestOne(string fmt, object[] values, int precision = 0)
         {
             var packed = PacketEncoding.Pack(fmt, values);
